requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
django==2.1.9             # via -r requirements.in, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
pytz==2019.3              # via django
